+++ src/Fieldtypes/Fieldtype.php
@@ -7,6 +7,10 @@ use Expressionengine\Coilpack\Models\FieldContent;
 
 abstract class Fieldtype
 {
+
+    public $name;
+    public $id;
+
     public function __construct(string $name, $id = null)
     {
         $this->name = $name;
@@ -26,6 +30,14 @@ abstract class Fieldtype
         return [];
     }
 
+    /**
+     * Call a modifier on the given field content
+     *
+     * @param FieldContent $content
+     * @param string $name
+     * @param array $parameters
+     * @return FieldtypeOutput
+     */
     public function callModifier(FieldContent $content, string $name, array $parameters = [])
     {
         if (array_key_exists($name, $this->modifiers()) && $this->modifiers()[$name] instanceof Modifier) {